requirements.txt
mcp>=1.0.0
httpx>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
cryptography>=42.0.0
langchain>=0.2.0
langchain-openai>=0.1.0
crewai>=0.30.0
openai>=1.30.0
